# Conflicts:
#	CMakeLists.txt
#	lib/fetch_and_build_solvers.sh
#	lib/fetch_solvers_with_curl.sh
#	scripts/server/create_mallob_env.sh
#	scripts/slurm/postrun.sh
#	scripts/slurm/sbatch-plain-satwp.sh
#	src/app/app_registry.cpp
#	src/app/sat/execution/engine.cpp
#	src/app/sat/job/anytime_sat_clause_communicator.cpp
#	src/app/sat/job/anytime_sat_clause_communicator.hpp
#	src/app/sat/options.hpp
#	src/app/sat/solvers/cadical.cpp
#	src/app/sat/solvers/kissat.cpp
#	src/app/sat/solvers/kissat.hpp
#	src/app/satwithpre/options.hpp
#	src/app/satwithpre/sat_preprocess_solver.hpp
#	src/app/satwithpre/sat_preprocessor.hpp
#	src/interface/json_interface.cpp

+++ scripts/slurm/postrun.sh
@@ -1,58 +1,42 @@
 #!/bin/bash
 
-source $ACCOUNTINFO # $projectname , $username
-
-echo "$username"
-echo "$projectname"
+source scripts/slurm/account.sh # $projectname , $username
 
 set -e
 
-if [ -z "${1:-}" ]; then
-	echo "usage: $0 <JOBNAME>"
-	exit 1
+jobname="$1"
+outdir="/hppfs/work/$projectname/$username/logs/${jobname}"
+
+openids=$(cat sbatch/generated/${jobname}/.remaining_ids)
+if ! [ -z "$openids" ]; then
+    echo "ERROR: Remaining open job IDs: $openids"
+    echo "If you are fine with this, execute \"> sbatch/generated/${jobname}/.remaining_ids\" and retry."
+    exit 1
 fi
 
-jobname="$1"
-
-echo "$jobname"
-# WORK
-outdir="/hppfs/work/$projectname/$username/logs/${jobname}"
-# HOME
-# outdir="$HOME/logs/${jobname}"
-
-echo "$jobname: moving results to unified $outdir"
-
 mkdir -p "$outdir/"
-# for f in /hppfs/work/$projectname/$username/logs/${jobname}-*/*/.alldone ; do
-for f in $outdir-*/*/.alldone ; do
+for f in /hppfs/work/$projectname/$username/logs/${jobname}-*/*/.alldone ; do
     if [ -d "$outdir/$(basename $(dirname $f))" ]; then continue; fi
-    # echo "mv $(dirname $f) $outdir/ "
     mv $(dirname $f) "$outdir/"
 done
-
-
-echo "mv sbatch/generated/${jobname}/sbatch.sh $outdir/"
 mv sbatch/generated/${jobname}/sbatch.sh "$outdir/"
-
-echo "$jobname: moving slurm-out files"
-echo $outdir-*/ | grep -oE "\-[0-9]{7}/" | grep -oE "[0-9]{7}" | while read slurmid; do
-    # echo "mv slurm-${slurmid}.out $outdir/"
+echo /hppfs/work/$projectname/$username/logs/${jobname}-*/ | grep -oE "\-[0-9]{7}/" | grep -oE "[0-9]{7}" | while read slurmid; do
+    if ! [ -f slurm-${slurmid}.out ]; then
+       echo "WARN: Slurm file slurm-${slurmid}.out not present"
+       continue
+    fi
     mv slurm-${slurmid}.out "$outdir/"
 done
 
 for globallogdir in $outdir/*/ ; do
     grep -m 1 "Program options" "$globallogdir/0/log.0" | grep -oP "mono=.*? " | sed 's/mono=//g' | awk '{print $1}' > "$globallogdir/instance.txt"
-    # echo "written $globallogdir/instance.txt  with $(cat $globallogdir/instance.txt)"
 done
 
 echo "All logs and sbatch / SLURM files moved to: $outdir"
 echo ""
 
-
-
-echo "$jobname: removing generated sbatch"
-rmdir sbatch/generated/${jobname}/.{done,reserved}* # locks for starting jobs
 rm sbatch/generated/${jobname}/.ticks # tick list for job counting failsafe
+rm sbatch/generated/${jobname}/.remaining_ids
 rmdir sbatch/generated/${jobname}
 
 echo ""
